feat(frontend): moving logout logic to utils folder
